Add eigen and PETSC_DIR to fenics metapackage
Eigen is a build and runtime dependency for some FEniCS packages.
Previously Eigen was added as only a build dependency.
Now it is also included as a runtime dependency.
In the future, this may be moved to fenics-instant, as I think
that that is what actually needs Eigen at runtime.

PETSC_DIR was not previously exported.
To fix this, a new activation and deactivation script was added to
the fenics metapackage.
Now PETSC_DIR is exported automatically.
In the future, this should be moved to the PETSc build itself.
This will be done the next time PETSc needs to be rebuilt.